fix: errors for secret client
